rename packege name from stream to messaging
