+++ api/src/main/java/com/datastrato/gravitino/messaging/TopicChange.java
@@ -2,7 +2,7 @@
  * Copyright 2024 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.stream;
+package com.datastrato.gravitino.messaging;
 
 import java.util.Objects;
